requirements.txt
python-dotenv

loguru

agno

google-genai
genai_processors

yfinance

mcp

googlesearch-python
ddgs
pycountry
baidusearch

curl-cffi==0.8.1b9
wasmtime
numpy
nodriver
drissionpage
fastapi
uvicorn
pyvirtualdisplay

# pip install droidrun[google] -e utils/droidrun/.

# ~/apps/illm/utils/EvoAgentX/Wonderful_workflow_corpus/invest/stock_analysis.py
# pip install -r  ~/apps/illm/utils/EvoAgentX/requirements.txt
evoagentx
matplotlib
akshare
telethon

# Reflex project
reflex

# ibacktest
backtesting
pandas
numpy

# Stock API
tushare
akshare

# Data analysis and visualization
matplotlib
seaborn
plotly

# Technical indicators
ta

# Utilities
python-dateutil
requests
pyyaml

# Optional: Jupyter support
jupyter
ipywidgets

reflex==0.8.15